+++ changelog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on http://keepachangelog.com/en/1.0.0/
 and this project adheres to http://semver.org/spec/v2.0.0.html.
 
-## [unreleased]
+## [0.28.0] 2024-09-21
 
 - Fix Subscribe.subscriptionID is optional
 - Add func Unsubscribe.Filters